requirements.txt
absl-py==2.1.0
attrs==24.2.0
cffi==1.17.1
contourpy==1.3.0
cycler==0.12.1
flatbuffers==24.3.25
fonttools==4.53.1
jax==0.4.32
jaxlib==0.4.32
kiwisolver==1.4.7
matplotlib==3.9.2
mediapipe==0.10.14
ml-dtypes==0.4.0
numpy==2.1.1
opencv-contrib-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pillow==10.4.0
protobuf==4.25.4
pycparser==2.22
pyparsing==3.1.4
python-dateutil==2.9.0.post0
scipy==1.14.1
six==1.16.0
sounddevice==0.5.0
